# Conflicts:
#	camel-core/readme-eip.adoc
#	camel-core/src/main/docs/content-based-router.adoc
#	camel-core/src/main/docs/eips/choice-eip.mv
#	camel-core/src/main/docs/eips/content-based-router.adoc
#	camel-core/src/main/docs/eips/filter-eip.adoc
#	camel-core/src/main/docs/eips/message-filter.adoc
#	camel-core/src/main/docs/message-filter.adoc

+++ platforms/spring-boot/components-starter/camel-salesforce-starter/src/main/java/org/apache/camel/component/salesforce/springboot/SalesforceComponentConfiguration.java
@@ -574,7 +574,7 @@ public class SalesforceComponentConfiguration {
          * 
          * @param defaultReplayId
          */
-        private Integer defaultReplayId;
+        private Long defaultReplayId;
         /**
          * Replay IDs to start from per channel name.
          */
@@ -909,11 +909,11 @@ public class SalesforceComponentConfiguration {
             this.maxBackoff = maxBackoff;
         }
 
-        public Integer getDefaultReplayId() {
+        public Long getDefaultReplayId() {
             return defaultReplayId;
         }
 
-        public void setDefaultReplayId(Integer defaultReplayId) {
+        public void setDefaultReplayId(Long defaultReplayId) {
             this.defaultReplayId = defaultReplayId;
         }
